Changes in code as per PR suggestions
