+++ src/main/java/ca/uhn/fhir/jpa/starter/JpaRestfulServer.java
@@ -1,5 +1,7 @@
 package ca.uhn.fhir.jpa.starter;
 
+import javax.servlet.ServletException;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
